Refactor dashboard logic and centralize auth handling
Moved auction and product dashboard logic into dedicated components for better separation of concerns. Introduced a new useAuth hook to centralize authentication and user role management, simplifying Header and page components. Updated routing and rendering logic in main pages to use user roles for conditional dashboard display. Cleaned up and modularized styles and code for maintainability.
