+++ LeafBidWeb/src/app/auction/view/page.module.css
@@ -1,0 +1,30 @@
+.main {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 40px;
+    gap: 32px;
+}
+
+.page {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    max-width: 80%;
+}
+
+.huidigeVeilingen {
+    text-align: left;
+    width: 100%;
+}
+
+.panels {
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+    width: 100%;
+}
+
+.padding{
+    padding-top: 24px;
+}
