+++ LeafBidWeb/src/components/Pages/AuctionView.module.css
